@@ -30,7 +30,7 @@
     "extract-text-webpack-plugin": "^1.0.1",
     "file-loader": "^0.8.5",
     "front-matter": "^2.0.7",
-    "fs-extra": "^0.28.0",
+    "fs-extra": "^0.30.0",
     "glob": "^7.0.0",
     "hapi": "^8.5.1",
     "hapi-webpack-plugin": "^1.3.0",
@@ -53,18 +53,18 @@
     "null-loader": "^0.1.1",
     "object-assign": "^4.0.1",
     "opn": "^4.0.1",
-    "parse-filepath": "^0.6.3",
+    "parse-filepath": "^1.0.1",
     "postcss-browser-reporter": "^0.4.0",
     "postcss-cssnext": "^2.5.2",
     "postcss-import": "^8.1.0",
-    "postcss-loader": "^0.8.1",
+    "postcss-loader": "^0.9.1",
     "postcss-reporter": "^1.3.3",
     "raw-loader": "^0.5.1",
-    "react": "^15.0.1",
+    "react": "^15.0.2",
     "react-document-title": "^2.0.1",
-    "react-dom": "^15.0.1",
+    "react-dom": "^15.0.2",
     "react-hot-loader": "^1.3.0",
-    "react-router": "^2.2.4",
+    "react-router": "^2.4.0",
     "sass-loader": "^3.1.2",
     "scroll-behavior": "^0.4.0",
     "slash": "^1.0.0",
@@ -87,10 +87,12 @@
     "babel-register": "^6.7.2",
     "bluebird": "^3.3.5",
     "cheerio": "^0.20.0",
-    "eslint": "^2.8.0",
-    "eslint-config-airbnb": "^6.1.0",
+    "eslint": "^2.9.0",
+    "eslint-config-airbnb": "^8.0.0",
     "eslint-plugin-ava": "^2.2.1",
-    "eslint-plugin-react": "^4.3.0",
+    "eslint-plugin-import": "^1.6.1",
+    "eslint-plugin-jsx-a11y": "^1.0.4",
+    "eslint-plugin-react": "^5.0.1",
     "nyc": "^6.4.0"
   },
   "engines": {
