@@ -1,4 +1,4 @@
-const exec = require('child_process').exec
+import { exec } from 'child_process'
 import fs from 'fs-extra'
 import sysPath from 'path'
 
@@ -73,7 +73,7 @@ const copy = (starterPath, rootPath, callback) => {
 //
 // Returns nothing.
 const clone = (address, rootPath, callback) => {
-  const gitHubRe = /(gh|github)\:(?:\/\/)?/
+  const gitHubRe = /(gh|github):(?:\/\/)?/
   const url = gitHubRe.test(address) ?
     (`git://github.com/${address.replace(gitHubRe, '')}.git`) : address
   logger.log(`Cloning git repo ${url} to ${rootPath}...`)
