@@ -20,12 +20,12 @@ export default function HTML (props) {
           name="viewport"
           content="width=device-width, initial-scale=1.0 maximum-scale=1.0"
         />
-        </head>
-        <body>
-          <div id="react-mount" dangerouslySetInnerHTML={{ __html: props.body }} />
-          <script src={prefixLink('/bundle.js')} />
-        </body>
-      </html>
+      </head>
+      <body>
+        <div id="react-mount" dangerouslySetInnerHTML={{ __html: props.body }} />
+        <script src={prefixLink('/bundle.js')} />
+      </body>
+    </html>
   )
 }
 
