@@ -2,7 +2,7 @@ import React from 'react'
 import { renderToString, renderToStaticMarkup } from 'react-dom/server'
 import { match, RouterContext } from 'react-router'
 import createRoutes from 'create-routes'
-import HTML from 'html'
+import Html from 'html'
 const loadContext = require('.gatsby-context')
 import { pages, config } from 'config'
 import { prefixLink } from '../isomorphic/gatsby-helpers'
@@ -29,7 +29,7 @@ module.exports = (locals, callback) => {
       }
 
       const html = `<!DOCTYPE html>\n ${renderToStaticMarkup(
-        <HTML body={body} {...renderProps} />
+        <Html body={body} {...renderProps} />
       )}`
       callback(null, html)
     }
